show big error message if no default props are specified
