+++ packages/cli/src/editor/components/RenderModal/SchemaEditor/SchemaEditor.tsx
@@ -1,45 +1,13 @@
 import React, {useCallback} from 'react';
 import {z} from 'remotion';
-import {Button} from '../../../../preview-server/error-overlay/remotion-overlay/Button';
-import {LIGHT_TEXT} from '../../../helpers/colors';
-import {Spacing} from '../../layout';
 import {VERTICAL_SCROLLBAR_CLASSNAME} from '../../Menu/is-menu-item';
-import {SchemaEmptyStateGraphic} from './SchemaEmptyStateGraphic';
-import {ZodErrorMessages} from './ZodErrorMessages';
+import {InvalidDefaultProps, InvalidSchema} from './SchemaErrorMessages';
 import {ZodObjectEditor} from './ZodObjectEditor';
-
-const errorExplanation: React.CSSProperties = {
-	fontSize: 14,
-	color: LIGHT_TEXT,
-	fontFamily: 'sans-serif',
-	lineHeight: 1.5,
-};
-
-const codeSnippet: React.CSSProperties = {
-	fontSize: 14,
-	color: 'var(--blue)',
-	fontFamily: 'monospace',
-};
 
 const scrollable: React.CSSProperties = {
 	padding: '8px 12px',
 	display: 'flex',
 	flexDirection: 'column',
-	overflowY: 'auto',
-};
-
-const explainer: React.CSSProperties = {
-	display: 'flex',
-	flex: 1,
-	flexDirection: 'column',
-	padding: '0 12px',
-	justifyContent: 'center',
-	alignItems: 'center',
-	textAlign: 'center',
-};
-
-const errorContainer: React.CSSProperties = {
-	padding: '8px 12px',
 	overflowY: 'auto',
 };
 
@@ -70,92 +38,15 @@ export const SchemaEditor: React.FC<{
 		setValue(defaultProps);
 	}, [defaultProps, setValue]);
 
-	const openDocs = useCallback(() => {
-		window.open(
-			// TODO: Make sure to update this link when we release v4
-			'https://v4.remotion.dev/docs/parametrized-rendering#define-a-schema-'
-		);
-	}, []);
-
-	if (typeName === z.ZodFirstPartyTypeKind.ZodAny) {
-		return (
-			<div style={explainer}>
-				<SchemaEmptyStateGraphic />
-				<Spacing y={5} />
-				<div style={errorExplanation}>
-					Make the props of this composition interactively editable by adding a{' '}
-					<code style={codeSnippet}>schema</code> prop to the{' '}
-					<code style={codeSnippet}>{'<Composition>'}</code> component.
-				</div>
-				<Spacing y={2} block />
-				<Button onClick={openDocs}>Learn how</Button>
-			</div>
-		);
-	}
-
 	if (!zodValidationResult.success) {
-		if (defaultProps === undefined) {
-			return (
-				<div style={explainer}>
-					<div style={errorExplanation}>
-						The schema can not be edited because the{' '}
-						<code style={codeSnippet}>defaultProps</code> prop in the{' '}
-						<code style={codeSnippet}>{'<Composition>'}</code> does not exist.
-					</div>
-					<Spacing y={1} />
-					<div style={errorExplanation}>
-						Fix the schema by adding a{' '}
-						<code style={codeSnippet}>defaultProps</code> prop to your
-						composition.
-					</div>
-					<Spacing y={2} block />
-					<Button onClick={openDocs}>Learn more</Button>
-				</div>
-			);
-		}
-
 		const defaultPropsValid = schema.safeParse(defaultProps);
 
 		if (!defaultPropsValid.success) {
-			// TODO: Does not react to when schema is updated
-			return (
-				<div style={errorContainer}>
-					<div style={errorExplanation}>
-						The schema can not be edited because the{' '}
-						<code style={codeSnippet}>defaultProps</code> prop in the{' '}
-						<code style={codeSnippet}>{'<Composition>'}</code> is not valid:
-					</div>
-					<Spacing y={1} block />
-					<ZodErrorMessages zodValidationResult={zodValidationResult} />
-					<Spacing y={1} block />
-					<div style={errorExplanation}>
-						Fix the schema by changing the{' '}
-						<code style={codeSnippet}>defaultProps</code> prop in your
-						composition so it does not give a type error.
-					</div>
-				</div>
-			);
+			return <InvalidDefaultProps zodValidationResult={zodValidationResult} />;
 		}
 
 		return (
-			<div style={errorContainer}>
-				<div style={errorExplanation}>
-					The data does not satisfy the schema:
-				</div>
-				<Spacing y={1} block />
-				<ZodErrorMessages zodValidationResult={zodValidationResult} />
-				<Spacing y={1} block />
-				<div style={errorExplanation}>
-					Fix the schema using the JSON editor.
-				</div>
-				<Spacing y={1} block />
-				<div style={errorExplanation}>
-					Alternatively, reset the data to the <code>defaultProps</code> that
-					you have defined.
-				</div>
-				<Spacing y={1} block />
-				<Button onClick={reset}>Reset props</Button>
-			</div>
+			<InvalidSchema reset={reset} zodValidationResult={zodValidationResult} />
 		);
 	}
